Refactor type system: expression types, explicit casts, dispatch checking
- Add dtype (ScalarType) to all expression IR nodes; ir_type_annotate
  now annotates every expression, not just IRAssign._resolved_type
- Replace per-codegen type inference heuristics (~40 lines each) with
  node.dtype lookups in CUDA, MSL, OpenCL backends
- Add explicit cast functions: pgc.f32/f64/i32/i64/u32/u64 in kernels;
  IRCast.dtype is now a ScalarType instead of "int"/"float" strings
- Float scalar args auto-promote to f64 when any field arg uses f64
- Add pgc.shared_like(field, size) — shared memory inherits field dtype
- Add dispatch-time type checking: validates field dtypes against backend
  capabilities (e.g., Metal rejects f64) with clear error messages
- LLVM backend handles full cast matrix (fpext, fptrunc, zext, sext, etc.)
- Update users guide, developers guide, CLAUDE.md, and examples
- Add 96 new tests covering all 5 layers of the refactor

+++ tests/test_type_annotate.py
@@ -57,7 +57,7 @@ def test_assign_from_binop_promotes():
 
 
 def test_assign_from_cast_int():
-    cast = ir.IRCast(value=ir.IRConstant(3.14), dtype="int")
+    cast = ir.IRCast(value=ir.IRConstant(3.14), dtype=i32)
     assign = ir.IRAssign(target="idx", value=cast)
     func = _make_func([], [assign])
     annotate_types(func)
@@ -65,7 +65,7 @@ def test_assign_from_cast_int():
 
 
 def test_assign_from_cast_float():
-    cast = ir.IRCast(value=ir.IRConstant(42), dtype="float")
+    cast = ir.IRCast(value=ir.IRConstant(42), dtype=f32)
     assign = ir.IRAssign(target="val", value=cast)
     func = _make_func([], [assign])
     annotate_types(func)
